Revert "Merge branch 'kk_2.7_rb1.17' of git://codeaurora.org/platform/packages/apps/Bluetooth into cm-11.0"
This reverts commit 4e438a0d5e0be9b92915d5380cfa8b2860c35589.

+++ src/com/android/bluetooth/opp/BluetoothOppTransferInfo.java
@@ -41,9 +41,9 @@ public class BluetoothOppTransferInfo {
 
     int mDirection;
 
-    long mTotalBytes;
+    int mTotalBytes;
 
-    long mCurrentBytes;
+    int mCurrentBytes;
 
     int mStatus;
